+++ base/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.2.1 (WIP)
+
+- Implement `RootRem` for `u8`, `u16`, `u32`
+- Add trait `Root` and implement it for `u8`, `u16`, `u32`, `u64`, `u128`
+
 ## 0.2.0
 
 - Add traits `Approximation`, `Sign` and `EstimatedLog2`.